django-scrapy-template/pyproject.toml
[project]
name = "django-scrapy-tracker"
version = "0.1.0"
description = "Django + Scrapy data tracking application"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16",
    "scrapy>=2.14",
    "scrapy-playwright>=0.0.46",
    "psycopg2-binary>=2.9",
    "django-filter>=25.0",
    "celery>=5.5",
    "django-celery-beat>=2.7",
    "redis>=7.0",
    "drf-spectacular>=0.29",
    "django-extensions>=4.0",
    "django-debug-toolbar>=6.0",
    "python-decouple>=3.8",
    "Pillow>=12.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-django>=4.10",
    "factory-boy>=3.3",
    "pytest-cov>=5.0",
    "black>=26.0",
    "isort>=6.0",
    "flake8>=7.3",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--tb=short"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
known_django = ["django"]
known_third_party = ["rest_framework", "scrapy", "celery"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
